// interfaces/onboarding/clients.ts
export interface ClientsFormData {
  company: string
  start_year: string
  street_address: string
  postal_code: string
  city: string
  state: string
  country: string
  vat_id: string
  contact_name: string
  email: string
  contact_number: string
  whatsapp_number: string
  website: string
  facebook: string
  instagram: string
  twitter: string
  google_business_profile_link: string
}